+++ src/avgpool_layer_kernels.cu
@@ -45,7 +45,7 @@ __global__ void backward_avgpool_layer_kernel(int n, int w, int h, int c, float 
 
 extern "C" void forward_avgpool_layer_gpu(avgpool_layer layer, network net)
 {
-    size_t n = layer.c*layer.batch;
+    uint64_t n = layer.c*layer.batch;
 
     forward_avgpool_layer_kernel<<<cuda_gridsize(n), BLOCK>>>(n, layer.w, layer.h, layer.c, net.input_gpu, layer.output_gpu);
     check_error(cudaPeekAtLastError());
@@ -53,7 +53,7 @@ extern "C" void forward_avgpool_layer_gpu(avgpool_layer layer, network net)
 
 extern "C" void backward_avgpool_layer_gpu(avgpool_layer layer, network net)
 {
-    size_t n = layer.c*layer.batch;
+    uint64_t n = layer.c*layer.batch;
 
     backward_avgpool_layer_kernel<<<cuda_gridsize(n), BLOCK>>>(n, layer.w, layer.h, layer.c, net.delta_gpu, layer.delta_gpu);
     check_error(cudaPeekAtLastError());
