+++ src/maxpool_layer_kernels.cu
@@ -90,7 +90,7 @@ extern "C" void forward_maxpool_layer_gpu(maxpool_layer layer, network net)
     int w = layer.out_w;
     int c = layer.c;
 
-    size_t n = h*w*c*layer.batch;
+    uint64_t n = h*w*c*layer.batch;
 
     forward_maxpool_layer_kernel<<<cuda_gridsize(n), BLOCK>>>(n, layer.h, layer.w, layer.c, layer.stride, layer.size, layer.pad, net.input_gpu, layer.output_gpu, layer.indexes_gpu);
     check_error(cudaPeekAtLastError());
@@ -98,7 +98,7 @@ extern "C" void forward_maxpool_layer_gpu(maxpool_layer layer, network net)
 
 extern "C" void backward_maxpool_layer_gpu(maxpool_layer layer, network net)
 {
-    size_t n = layer.h*layer.w*layer.c*layer.batch;
+    uint64_t n = layer.h*layer.w*layer.c*layer.batch;
 
     backward_maxpool_layer_kernel<<<cuda_gridsize(n), BLOCK>>>(n, layer.h, layer.w, layer.c, layer.stride, layer.size, layer.pad, layer.delta_gpu, net.delta_gpu, layer.indexes_gpu);
     check_error(cudaPeekAtLastError());
